Added AskForPermissionsConsent card
Made the Card class abstract, and moved the type specific properties to their respective card classes. Also added the AskForPermissionsConsent card to enable the Address API for example.
